+++ packages/gatsby/src/bootstrap/schema-hot-reloader.ts
@@ -1,27 +1,38 @@
-const { debounce, cloneDeep } = require(`lodash`)
-const { emitter, store } = require(`../redux`)
-const { rebuild } = require(`../schema`)
-const { haveEqualFields } = require(`../schema/infer/inference-metadata`)
-const { updateStateAndRunQueries } = require(`../query/query-watcher`)
-const report = require(`gatsby-cli/lib/reporter`)
+import { debounce, cloneDeep } from "lodash"
+import { emitter, store } from "../redux"
+import { rebuild } from "../schema"
+import { haveEqualFields } from "../schema/infer/inference-metadata"
+import { updateStateAndRunQueries } from "../query/query-watcher"
+import report from "gatsby-cli/lib/reporter"
+import { IGatsbyState } from "../redux/types"
 
-const inferredTypesChanged = (typeMap, prevTypeMap) =>
+type TypeMap = IGatsbyState["inferenceMetadata"]["typeMap"]
+type SchemaCustomization = IGatsbyState["schemaCustomization"]
+type InferenceMetadata = IGatsbyState["inferenceMetadata"]
+
+const inferredTypesChanged = (
+  typeMap: TypeMap,
+  prevTypeMap: TypeMap
+): boolean =>
   Object.keys(typeMap).some(
     type =>
       typeMap[type].dirty && !haveEqualFields(typeMap[type], prevTypeMap[type])
   )
 
-const schemaChanged = (schemaCustomization, lastSchemaCustomization) =>
+const schemaChanged = (
+  schemaCustomization: SchemaCustomization,
+  lastSchemaCustomization: SchemaCustomization
+): boolean =>
   [`fieldExtensions`, `printConfig`, `thirdPartySchemas`, `types`].some(
     key => schemaCustomization[key] !== lastSchemaCustomization[key]
   )
 
-let lastMetadata
-let lastSchemaCustomization
+let lastMetadata: InferenceMetadata
+let lastSchemaCustomization: SchemaCustomization
 
 // API_RUNNING_QUEUE_EMPTY could be emitted multiple types
 // in a short period of time, so debounce seems reasonable
-const maybeRebuildSchema = debounce(async () => {
+const maybeRebuildSchema = debounce(async (): Promise<void> => {
   const { inferenceMetadata, schemaCustomization } = store.getState()
 
   if (
@@ -40,7 +51,7 @@ const maybeRebuildSchema = debounce(async () => {
   activity.end()
 }, 1000)
 
-module.exports = () => {
+export const bootstrapSchemaHotReloader = (): void => {
   const { inferenceMetadata, schemaCustomization } = store.getState()
   lastMetadata = cloneDeep(inferenceMetadata)
   lastSchemaCustomization = schemaCustomization