+++ packages/gatsby/src/query/queue.js
@@ -7,7 +7,8 @@ const GraphQLRunner = require(`./graphql-runner`).default
 
 const createBaseOptions = () => {
   return {
-    concurrent: 4,
+    concurrent: Number(process.env.GATSBY_EXPERIMENTAL_QUERY_CONCURRENCY) || 4,
+    // eslint-disable-next-line new-cap
     store: FastMemoryStore(),
   }
 }
@@ -70,8 +71,7 @@ const createDevelopQueue = getRunner => {
  * Note: queue is reused in develop so make sure to thoroughly cleanup hooks
  */
 const processBatch = async (queue, jobs, activity) => {
-  const numJobs = jobs.length
-  if (numJobs === 0) {
+  if (jobs.length === 0) {
     return Promise.resolve()
   }