+++ packages/gatsby/src/redux/actions/internal.ts
@@ -16,10 +16,6 @@ import {
 /**
  * Create a dependency between a page and data. Probably for
  * internal use only.
- * @param {Object} $0
- * @param {string} $0.path the path to the page
- * @param {string} $0.nodeId A node ID
- * @param {string} $0.connection A connection type
  * @private
  */
 export const createPageDependency = (
@@ -27,7 +23,7 @@ export const createPageDependency = (
     path,
     nodeId,
     connection,
-  }: { path: string; nodeId: string; connection: string },
+  }: { path: string; nodeId?: string; connection?: string },
   plugin = ``
 ): ICreatePageDependencyAction => {
   return {
@@ -44,7 +40,6 @@ export const createPageDependency = (
 /**
  * Delete dependencies between an array of pages and data. Probably for
  * internal use only. Used when deleting pages.
- * @param {Array} paths the paths to delete.
  * @private
  */
 export const deleteComponentsDependencies = (
@@ -99,11 +94,6 @@ export const replaceStaticQuery = (
  *
  * Report that a query has been extracted from a component. Used by
  * query-compiler.js.
- *
- * @param {Object} $0
- * @param {componentPath} $0.componentPath The path to the component that just had
- * its query read.
- * @param {query} $0.query The GraphQL query that was extracted from the component.
  * @private
  */
 export const queryExtracted = (
@@ -122,11 +112,6 @@ export const queryExtracted = (
 /**
  *
  * Report that the Relay Compiler found a graphql error when attempting to extract a query
- *
- * @param {Object} $0
- * @param {componentPath} $0.componentPath The path to the component that just had
- * its query read.
- * @param {error} $0.error The GraphQL query that was extracted from the component.
  * @private
  */
 export const queryExtractionGraphQLError = (
@@ -146,10 +131,6 @@ export const queryExtractionGraphQLError = (
  *
  * Report that babel was able to extract the graphql query.
  * Indicates that the file is free of JS errors.
- *
- * @param {Object} $0
- * @param {componentPath} $0.componentPath The path to the component that just had
- * its query read.
  * @private
  */
 export const queryExtractedBabelSuccess = (
@@ -168,11 +149,6 @@ export const queryExtractedBabelSuccess = (
 /**
  *
  * Report that the Relay Compiler found a babel error when attempting to extract a query
- *
- * @param {Object} $0
- * @param {componentPath} $0.componentPath The path to the component that just had
- * its query read.
- * @param {error} $0.error The Babel error object
  * @private
  */
 export const queryExtractionBabelError = (
@@ -190,8 +166,6 @@ export const queryExtractionBabelError = (
 
 /**
  * Set overall program status e.g. `BOOTSTRAPING` or `BOOTSTRAP_FINISHED`.
- *
- * @param {string} Program status
  * @private
  */
 export const setProgramStatus = (
@@ -209,8 +183,6 @@ export const setProgramStatus = (
 
 /**
  * Broadcast that a page's query was run.
- *
- * @param {string} Path to the page component that changed.
  * @private
  */
 export const pageQueryRun = (
@@ -228,8 +200,6 @@ export const pageQueryRun = (
 
 /**
  * Remove jobs which are marked as stale (inputPath doesn't exists)
- *
- * @param {string} contentDigest
  * @private
  */
 export const removeStaleJob = (