+++ packages/gatsby/src/redux/nodes.ts
@@ -1,14 +1,11 @@
-/* @flow */
-
+import { store } from "./"
+import { IGatsbyNode } from "./types"
 import { createPageDependency } from "./actions/add-page-dependency"
-const { store } = require(`./index`)
 
 /**
  * Get all nodes from redux store.
- *
- * @returns {Array}
  */
-const getNodes = () => {
+export const getNodes = (): IGatsbyNode[] => {
   const nodes = store.getState().nodes
   if (nodes) {
     return Array.from(nodes.values())
@@ -17,24 +14,16 @@ const getNodes = () => {
   }
 }
 
-exports.getNodes = getNodes
-
-/** Get node by id from store.
- *
- * @param {string} id
- * @returns {Object}
+/**
+ * Get node by id from store.
  */
-const getNode = id => store.getState().nodes.get(id)
-
-exports.getNode = getNode
+export const getNode = (id: string): IGatsbyNode | undefined =>
+  store.getState().nodes.get(id)
 
 /**
  * Get all nodes of type from redux store.
- *
- * @param {string} type
- * @returns {Array}
  */
-const getNodesByType = type => {
+export const getNodesByType = (type: string): IGatsbyNode[] => {
   const nodes = store.getState().nodesByType.get(type)
   if (nodes) {
     return Array.from(nodes.values())
@@ -43,25 +32,16 @@ const getNodesByType = type => {
   }
 }
 
-exports.getNodesByType = getNodesByType
-
 /**
  * Get all type names from redux store.
- *
- * @returns {Array}
  */
-const getTypes = () => Array.from(store.getState().nodesByType.keys())
-
-exports.getTypes = getTypes
+export const getTypes = (): string[] =>
+  Array.from(store.getState().nodesByType.keys())
 
 /**
  * Determine if node has changed.
- *
- * @param {string} id
- * @param {string} digest
- * @returns {boolean}
  */
-exports.hasNodeChanged = (id, digest) => {
+export const hasNodeChanged = (id: string, digest: string): boolean => {
   const node = store.getState().nodes.get(id)
   if (!node) {
     return true
@@ -72,53 +52,64 @@ exports.hasNodeChanged = (id, digest) => {
 
 /**
  * Get node and save path dependency.
- *
- * @param {string} id
- * @param {string} path
- * @returns {Object} node
  */
-exports.getNodeAndSavePathDependency = (id, path) => {
+export const getNodeAndSavePathDependency = (
+  id: string,
+  path: string
+): IGatsbyNode | undefined => {
   const node = getNode(id)
+
+  if (!node) {
+    console.error(
+      `getNodeAndSavePathDependency failed for node id: ${id} as it was not found in cache`
+    )
+    return undefined
+  }
+
   createPageDependency({ path, nodeId: id })
   return node
 }
 
-exports.saveResolvedNodes = async (nodeTypeNames, resolver) => {
+type Resolver = (node: IGatsbyNode) => Promise<any> // TODO
+
+export const saveResolvedNodes = async (
+  nodeTypeNames: string[],
+  resolver: Resolver
+): Promise<void> => {
   for (const typeName of nodeTypeNames) {
     const nodes = store.getState().nodesByType.get(typeName)
+    if (!nodes) return
+
     const resolvedNodes = new Map()
-    if (nodes) {
-      for (const node of nodes.values()) {
-        const resolved = await resolver(node)
-        resolvedNodes.set(node.id, resolved)
-      }
-      store.dispatch({
-        type: `SET_RESOLVED_NODES`,
-        payload: {
-          key: typeName,
-          nodes: resolvedNodes,
-        },
-      })
+    for (const node of nodes.values()) {
+      const resolved = await resolver(node)
+      resolvedNodes.set(node.id, resolved)
     }
+    store.dispatch({
+      type: `SET_RESOLVED_NODES`,
+      payload: {
+        key: typeName,
+        nodes: resolvedNodes,
+      },
+    })
   }
 }
 
 /**
  * Get node and save path dependency.
- *
- * @param {string} typeName
- * @param {string} id
- * @returns {Object|void} node
  */
-const getResolvedNode = (typeName, id) => {
+export const getResolvedNode = (
+  typeName: string,
+  id: string
+): IGatsbyNode | null => {
   const { nodesByType, resolvedNodesCache } = store.getState()
-  const nodes /*: Map<mixed> */ = nodesByType.get(typeName)
+  const nodes = nodesByType.get(typeName)
 
   if (!nodes) {
     return null
   }
 
-  let node = nodes.get(id)
+  const node = nodes.get(id)
 
   if (!node) {
     return null
@@ -133,27 +124,26 @@ const getResolvedNode = (typeName, id) => {
   return node
 }
 
-exports.getResolvedNode = getResolvedNode
-
-const addResolvedNodes = (typeName, arr) => {
+export const addResolvedNodes = (typeName: string): IGatsbyNode[] => {
+  const resolvedNodes: IGatsbyNode[] = []
   const { nodesByType, resolvedNodesCache } = store.getState()
-  const nodes /*: Map<mixed> */ = nodesByType.get(typeName)
+  const nodes = nodesByType.get(typeName)
 
   if (!nodes) {
-    return
+    return []
   }
 
-  const resolvedNodes = resolvedNodesCache.get(typeName)
+  const resolvedNodesFromCache = resolvedNodesCache.get(typeName)
 
   nodes.forEach(node => {
-    if (resolvedNodes) {
-      node.__gatsby_resolved = resolvedNodes.get(node.id)
+    if (resolvedNodesFromCache) {
+      node.__gatsby_resolved = resolvedNodesFromCache.get(node.id)
     }
-    arr.push(node)
+    resolvedNodes.push(node)
   })
-}
 
-exports.addResolvedNodes = addResolvedNodes
+  return resolvedNodes
+}
 
 /**
  * Given a ("flat") filter path leading up to "eq", a set of node types, and a
@@ -161,32 +151,28 @@ exports.addResolvedNodes = addResolvedNodes
  * all the Nodes in a Set (or, if the property is `id`, just the Nodes).
  * This cache is used for applying the filter and is a massive improvement over
  * looping over all the nodes, when the number of pages (/nodes) scale up.
- *
- * @param {Array<string>} chain
- * @param {Array<string>} nodeTypeNames
- * @param {undefined | Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
- *   This object lives in query/query-runner.js and is passed down runQuery
- * @returns {undefined}
  */
-const ensureIndexByTypedChain = (
-  chain,
-  nodeTypeNames,
-  typedKeyValueIndexes
-) => {
+export const ensureIndexByTypedChain = (
+  chain: string[],
+  nodeTypeNames: string[],
+  typedKeyValueIndexes: Map<
+    string,
+    Map<string | number | boolean, Set<IGatsbyNode>>
+  >
+): void => {
   const chained = chain.join(`+`)
 
   const nodeTypeNamePrefix = nodeTypeNames.join(`,`) + `/`
   // The format of the typedKey is `type,type/path+to+eqobj`
   const typedKey = nodeTypeNamePrefix + chained
 
-  let byKeyValue = typedKeyValueIndexes.get(typedKey)
-  if (byKeyValue) {
+  if (typedKeyValueIndexes.has(typedKey)) {
     return
   }
 
   const { nodes, resolvedNodesCache } = store.getState()
 
-  byKeyValue = new Map() // Map<node.value, Set<all nodes with this value for this key>>
+  const byKeyValue = new Map<string | number | boolean, Set<IGatsbyNode>>()
   typedKeyValueIndexes.set(typedKey, byKeyValue)
 
   nodes.forEach(node => {
@@ -201,7 +187,7 @@ const ensureIndexByTypedChain = (
       node.__gatsby_resolved = resolvedNodes?.get(node.id)
     }
 
-    let v = node
+    let v = node as any
     let i = 0
     while (i < chain.length && v) {
       const nextProp = chain[i++]
@@ -221,14 +207,6 @@ const ensureIndexByTypedChain = (
       return
     }
 
-    // Special case `id` as that bucket never receives more than one element
-    if (chained === `id`) {
-      // Note: this is not a duplicate from `nodes` because this set only
-      //       contains nodes of this type. Page nodes are a subset of all nodes
-      byKeyValue.set(v, node)
-      return
-    }
-
     let set = byKeyValue.get(v)
     if (!set) {
       set = new Set()
@@ -237,8 +215,6 @@ const ensureIndexByTypedChain = (
     set.add(node)
   })
 }
-
-exports.ensureIndexByTypedChain = ensureIndexByTypedChain
 
 /**
  * Given a ("flat") filter path leading up to "eq", a target value to filter
@@ -252,26 +228,17 @@ exports.ensureIndexByTypedChain = ensureIndexByTypedChain
  *
  * The only exception is `id`, since internally there can be at most one node
  * per `id` so there's a minor optimization for that (no need for Sets).
- *
- * @param {Array<string>} chain Note: `eq` is assumed to be the leaf prop here
- * @param {boolean | number | string} value This is the value being filtered for
- * @param {Array<string>} nodeTypeNames
- * @param {undefined | Map<string, Map<string | number | boolean, Node>>} typedKeyValueIndexes
- *   This object lives in query/query-runner.js and is passed down runQuery
- * @returns {Array<Node> | undefined}
  */
-const getNodesByTypedChain = (
-  chain,
-  value,
-  nodeTypeNames,
-  typedKeyValueIndexes
-) => {
+export const getNodesByTypedChain = (
+  chain: string[],
+  value: boolean | number | string,
+  nodeTypeNames: string[],
+  typedKeyValueIndexes: Map<string, Map<string | number | boolean, IGatsbyNode>>
+): IGatsbyNode | undefined => {
   const key = chain.join(`+`)
 
   const typedKey = nodeTypeNames.join(`,`) + `/` + key
 
-  let byTypedKey = typedKeyValueIndexes?.get(typedKey)
+  const byTypedKey = typedKeyValueIndexes?.get(typedKey)
   return byTypedKey?.get(value)
 }
-
-exports.getNodesByTypedChain = getNodesByTypedChain