+++ packages/gatsby/src/utils/babel-parse-to-ast.ts
@@ -1,12 +1,11 @@
-/* @flow */
-const parser = require(`@babel/parser`)
+import { parse, ParserOptions } from "@babel/parser"
+import { File } from "@babel/types"
 
-const PARSER_OPTIONS = {
+const PARSER_OPTIONS: ParserOptions = {
   allowImportExportEverywhere: true,
   allowReturnOutsideFunction: true,
   allowSuperOutsideMethod: true,
-  sourceType: `unambigious`,
-  sourceFilename: true,
+  sourceType: `unambiguous`,
   plugins: [
     `jsx`,
     `flow`,
@@ -43,13 +42,13 @@ const PARSER_OPTIONS = {
   ],
 }
 
-export function getBabelParserOptions(filePath: string) {
+export function getBabelParserOptions(filePath: string): ParserOptions {
   // Flow and TypeScript plugins can't be enabled simultaneously
   if (/\.tsx?/.test(filePath)) {
     const { plugins } = PARSER_OPTIONS
     return {
       ...PARSER_OPTIONS,
-      plugins: plugins.map(plugin =>
+      plugins: plugins!.map(plugin =>
         plugin === `flow` ? `typescript` : plugin
       ),
     }
@@ -57,6 +56,6 @@ export function getBabelParserOptions(filePath: string) {
   return PARSER_OPTIONS
 }
 
-export function babelParseToAst(contents: string, filePath: string) {
-  return parser.parse(contents, getBabelParserOptions(filePath))
+export function babelParseToAst(contents: string, filePath: string): File {
+  return parse(contents, getBabelParserOptions(filePath))
 }